+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 12.34.0
+
+* `[fsutil]` Added method `TouchFile` for creating empty files
+* `[fsutil]` Code refactoring
+* `[fmtc]` Documentation refactoring
+* `[timeutil]` Documentation refactoring
+* `[fsutil]` Added usage examples
+
 ### 12.33.0
 
 * `[errutil]` Added support for `string`, `[]string`, and `errutil.Errors` types to method `Errors.Add`